Code Refactor
- Organized the project to be more easily navigated. Folders > Classes.
- Renamed several classes so all in the project follow the same template.
- Removed unnecessary references from classes.
- Dropped the HarmonyLib nuget package and went with the physical dll instead since the nuget package kept misbehaving. -_-
