--- Source/CompProperties_FireOverlaySouth.cs
@@ -1,21 +1,0 @@
-﻿using RimWorld;
-using Verse;
-using Verse.AI;
-using UnityEngine;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DankPyon_MedievalOverhaul
-{
-    public class CompProperties_FireOverlaySouth : CompProperties_FireOverlay
-    {
-        public CompProperties_FireOverlaySouth()
-        {
-            compClass = typeof(CompFireOverlaySouth);
-        }
-
-    }
-}
