+++ Source/Harmony/HarmonyInstance.cs
@@ -4,12 +4,9 @@ using RimWorld;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UnityEngine;
 using Verse;
 using Verse.AI;
-using Verse.Noise;
 
 namespace DankPyon
 {
