+++ Source/Harmony/Patch_FollowAndMeleeAttack.cs
