+++ Source/Settings/MedievalOverhaulMod.cs
@@ -1,14 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Security.Cryptography;
-using System.Text;
-using System.Threading.Tasks;
-using System.Xml;
-using UnityEngine;
-using UnityEngine.Diagnostics;
+﻿using UnityEngine;
 using Verse;
-using Verse.Noise;
 
 namespace DankPyon
 {
