+++ Source/ThingClasses/Building_Lootable.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using UnityEngine;
 using RimWorld;
 using Verse;
