+++ Source/ThingClasses/Bullet_ExplosiveProperty.cs
@@ -1,5 +1,4 @@
 ﻿using RimWorld;
-using RimWorld.Planet;
 using Verse;
 
 namespace DankPyon
