+++ Source/ThingClasses/DefModExtensions/AdditionalProjectileProperties.cs
@@ -1,18 +1,12 @@
-﻿using RimWorld;
-using RimWorld.Planet;
-using Verse;
+﻿using Verse;
 
 namespace DankPyon
 {
 
     public class AdditionalProjectileProperties : DefModExtension
     {
-
         public static readonly AdditionalProjectileProperties defaultValues = new AdditionalProjectileProperties();
-
         public ProjectileProperties projectile2;
-
-
     }
 
 }
