+++ Source/ThingClasses/DefModExtensions/BuildingLootableExtension.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using UnityEngine;
 using Verse;
 
 namespace MedievalOverhaul
