+++ Source/ThingClasses/DefModExtensions/ButcherProperties.cs
@@ -1,5 +1,4 @@
 ﻿using Verse;
-using RimWorld;
 
 namespace DankPyon
 {
