+++ Source/ThingClasses/DefModExtensions/PlantExtension.cs
@@ -1,6 +1,4 @@
-﻿using RimWorld;
-using RimWorld.Planet;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Verse;
 
 namespace DankPyon
