+++ Source/ThingComps/CompProps/CompProperties_FireOverlaySouth.cs
@@ -1,0 +1,13 @@
+﻿using RimWorld;
+
+namespace DankPyon_MedievalOverhaul
+{
+    public class CompProperties_FireOverlaySouth : CompProperties_FireOverlay
+    {
+        public CompProperties_FireOverlaySouth()
+        {
+            compClass = typeof(Comp_FireOverlaySouth);
+        }
+
+    }
+}
