+++ Source/ThingComps/CompProps/CompProperties_PawnSpawnerOnDestroy.cs
@@ -1,7 +1,5 @@
 ﻿using RimWorld;
 using Verse;
-using System;
-using System.Collections.Generic;
 
 namespace MedievalOverhaul
 {
