+++ Source/ThingComps/Comp_PawnSpawnerOnDestroy.cs
@@ -1,6 +1,5 @@
 ﻿using RimWorld;
 using Verse;
-using System;
 using System.Collections.Generic;
 using Verse.Sound;
