+++ Source/Tools/Tool_Melee.cs
@@ -1,7 +1,4 @@
-﻿using System.Text;
-using System.Threading.Tasks;
-using UnityEngine;
-using Verse;
+﻿using Verse;
 
 namespace DankPyon
 {
